refactor: Move accessor into raw apis
