+++ src/adapters/kv/backend.rs
@@ -19,6 +19,7 @@ use futures::AsyncReadExt;
 use super::Adapter;
 use crate::ops::*;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::*;
 
 /// Backend of kv service.
