+++ src/io_util/walk.rs
@@ -21,6 +21,7 @@ use async_trait::async_trait;
 use crate::object::*;
 use crate::ops::OpList;
 use crate::path::normalize_path;
+use crate::raw::*;
 use crate::*;
 
 const WALK_BUFFER_SIZE: usize = 256;
