+++ src/layers/tracing.rs
@@ -29,6 +29,7 @@ use crate::object::BlockingObjectPager;
 use crate::object::ObjectPage;
 use crate::object::ObjectPager;
 use crate::ops::*;
+use crate::raw::*;
 use crate::*;
 
 /// TracingLayer will add tracing for OpenDAL.
