+++ src/operator.rs
@@ -21,9 +21,8 @@ use futures::TryStreamExt;
 use crate::io_util::BottomUpWalker;
 use crate::io_util::TopDownWalker;
 use crate::object::ObjectLister;
+use crate::raw::*;
 use crate::services;
-use crate::Accessor;
-use crate::AccessorMetadata;
 use crate::Error;
 use crate::ErrorKind;
 use crate::Layer;
