+++ src/services/azblob/backend.rs
@@ -30,14 +30,13 @@ use reqsign::AzureStorageSigner;
 
 use super::dir_stream::DirStream;
 use super::error::parse_error;
-use crate::accessor::AccessorCapability;
-use crate::accessor::AccessorMetadata;
 use crate::http_util::*;
 use crate::object::ObjectMetadata;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
