+++ src/services/gcs/backend.rs
@@ -34,12 +34,12 @@ use super::dir_stream::DirStream;
 use super::error::parse_error;
 use super::error::parse_json_deserialize_error;
 use super::uri::percent_encode_path;
-use crate::accessor::AccessorCapability;
 use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
