+++ src/services/hdfs/backend.rs
@@ -27,12 +27,12 @@ use time::OffsetDateTime;
 
 use super::dir_stream::DirStream;
 use super::error::parse_io_error;
-use crate::accessor::AccessorCapability;
 use crate::object::EmptyObjectPager;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_rooted_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
