+++ src/services/http/backend.rs
@@ -22,11 +22,11 @@ use http::StatusCode;
 use log::debug;
 
 use super::error::parse_error;
-use crate::accessor::AccessorCapability;
 use crate::http_util::*;
 use crate::ops::*;
 use crate::path::build_rooted_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
