+++ src/services/ipfs/backend.rs
@@ -25,13 +25,13 @@ use prost::Message;
 
 use super::error::parse_error;
 use super::ipld::PBNode;
-use crate::accessor::AccessorCapability;
 use crate::http_util::*;
 use crate::object::ObjectPage;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_rooted_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
