+++ src/services/ipmfs/builder.rs
@@ -17,8 +17,8 @@ use log::debug;
 use super::backend::Backend;
 use crate::http_util::HttpClient;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
-use crate::Accessor;
 use crate::Result;
 
 /// Builder for service ipfs.
