+++ src/services/oss/backend.rs
@@ -32,12 +32,12 @@ use reqsign::AliyunOssSigner;
 use super::dir_stream::DirStream;
 use super::error::parse_error;
 use super::uri::percent_encode_path_hard;
-use crate::accessor::AccessorCapability;
 use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
