+++ src/services/s3/backend.rs
@@ -40,12 +40,12 @@ use serde::Serialize;
 use super::dir_stream::DirStream;
 use super::error::parse_error;
 use super::error::parse_xml_deserialize_error;
-use crate::accessor::AccessorCapability;
 use crate::http_util::*;
 use crate::object::ObjectPager;
 use crate::ops::*;
 use crate::path::build_abs_path;
 use crate::path::normalize_root;
+use crate::raw::*;
 use crate::wrappers::wrapper;
 use crate::*;
